# Fix keep EXIF data -ke / --keep-exif with BytesIO buffer