@YellowTech 2024-05-25
Can we merge this? If this gets through I could work on HEIC support with pillow-heif.