@victordomingos 2024-05-25
Done. I will perform some testing before issuing the next release.m, so it may take a while before it comes to pip.